module all test modules to support dir
