+++ test/elixir_ast_dialect_test.exs
@@ -4,28 +4,7 @@ defmodule ELXDialectTest do
   @moduletag :smoke
 
   test "gen elx from ast", %{ctx: ctx} do
-    ast =
-      quote do
-        defmodule TwoFuncMod do
-          def add_two_literal() do
-            a = 1 + 2
-            b = a + 3
-            b
-          end
-
-          def main() do
-            add_two_literal()
-          end
-        end
-
-        defmodule OneFuncMod do
-          def add_two_literal() do
-            a = 1 + 2
-            b = a + 3
-            b
-          end
-        end
-      end
+    ast = ElixirAST.Sample.naive_modules()
 
     mlir_module =
       ast
