+++ test/support/pass_raising_exception.ex
@@ -1,0 +1,8 @@
+defmodule PassRaisingException do
+  @moduledoc false
+  use Beaver.MLIR.Pass, on: "func.func"
+
+  def run(_op, _state) do
+    raise "exception in pass run"
+  end
+end
